Upgraded to Discord packages v1.0.2
WORKING
This fixes the problem that was caused when Discord added channel categories. (The problem was that the bot didn't work anymore).

IMPORTANT
You need to add the following to your Nuget Package sources:
https://www.myget.org/F/discord-net/api/v3/index.json

THEN, after adding the above source, you need to update all Discord packages to the latest pre-release version. (Apparently, the pre-release is read, but the developers that maintain that feed are very lazy.)